* Moved ExecuteRestRequest from AbiquoClient v1 to BaseAbiquoClient * Move AbiquoClientFactory into factory folder
