+++ src/biz.dfch.CS.Abiquo.Client.Tests/biz.dfch.CS.Abiquo.Client.Tests.csproj
@@ -193,7 +193,7 @@
     <Compile Include="..\AssemblyVersion.cs">
       <Link>Properties\AssemblyVersion.cs</Link>
     </Compile>
-    <Compile Include="AbiquoClientFactoryTest.cs" />
+    <Compile Include="Factory\AbiquoClientFactoryTest.cs" />
     <Compile Include="Authentication\BasicAuthenticationInformationTest.cs" />
     <Compile Include="IntegrationTestEnvironment.cs" />
     <Compile Include="v1\AbiquoClientIntegrationTests.cs" />
